# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "AppealDecisionIn": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "uphold",
              "reverse"
            ],
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "maxLength": 1000,
            "minLength": 2,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "note"
        ],
        "title": "AppealDecisionIn",
        "type": "object"
      },
...
      "AppealItem": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionOut"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "nickname",
          "text",
          "created_at",
          "action"
        ],
        "title": "AppealItem",
        "type": "object"
      },
...
      "DecisionIn": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "dismiss",
              "hide",
              "remove",
              "restore",
              "none"
            ],
            "title": "Decision",
            "type": "string"
          },
          "ground": {
            "maxLength": 40,
            "minLength": 2,
            "title": "Ground",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "sanction": {
            "default": "none",
            "enum": [
              "none",
              "auto",
              "warn",
              "limit_posting",
              "ban"
            ],
            "title": "Sanction",
            "type": "string"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "enum": [
              "post",
              "profile",
              "link"
            ],
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "target_type",
          "target_id",
          "decision",
          "ground"
        ],
        "title": "DecisionIn",
        "type": "object"
      },
      "DecisionOut": {
        "properties": {
          "actions": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Actions",
            "type": "array"
          },
          "resolved_reports": {
            "title": "Resolved Reports",
            "type": "integer"
          },
          "sanction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sanction"
          }
        },
        "required": [
          "actions",
          "sanction",
          "resolved_reports"
        ],
        "title": "DecisionOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SanctionIn": {
        "additionalProperties": false,
        "properties": {
          "ground": {
            "maxLength": 40,
            "minLength": 2,
            "title": "Ground",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "sanction": {
            "enum": [
              "warn",
              "limit_posting",
              "ban"
            ],
            "title": "Sanction",
            "type": "string"
          }
        },
        "required": [
          "sanction",
          "ground"
        ],
        "title": "SanctionIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WearX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/appeals/{appeal_id}": {
      "post": {
        "operationId": "decide_appeal_v1_admin_appeals__appeal_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "appeal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Appeal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppealDecisionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Appeal",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/reports/decide": {
      "post": {
        "operationId": "decide_v1_admin_reports_decide_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/reports/queue": {
      "get": {
        "operationId": "queue_v1_admin_reports_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QueueItem"
                  },
                  "title": "Response Queue V1 Admin Reports Queue Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/users/{nickname}": {
      "get": {
        "operationId": "user_case_v1_admin_users__nickname__get",
        "parameters": [
          {
            "in": "path",
            "name": "nickname",
            "required": true,
            "schema": {
              "title": "Nickname",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCase"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "User Case",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/users/{nickname}/sanction": {
      "post": {
        "operationId": "sanction_v1_admin_users__nickname__sanction_post",
        "parameters": [
          {
            "in": "path",
            "name": "nickname",
            "required": true,
            "schema": {
              "title": "Nickname",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SanctionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sanction",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/users/{nickname}": {
      "get": {
        "operationId": "get_user_v1_users__nickname__get",
        "parameters": [
          {
            "in": "path",
            "name": "nickname",
            "required": true,
            "schema": {
              "title": "Nickname",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get User",
        "tags": [
          "portfolio"
        ]
      }
    },
...
  }
}